move _shared_blobs into utils to remove the import cycle
