+++ src/huggingface_hub/constants.py
@@ -342,7 +342,7 @@ HF_HUB_DISABLE_XET: bool = _is_true(os.environ.get("HF_HUB_DISABLE_XET"))
 # Disable the cache-wide shared blob store.
 # Xet-backed files are stored in `<cache_dir>/blobs/<prefix>/<xet_hash>` and symlinked
 # into repo caches so identical files are stored only once across repos. See
-# `huggingface_hub._shared_blobs`.
+# `huggingface_hub.utils._shared_blobs`.
 HF_HUB_DISABLE_SHARED_BLOBS: bool = _is_true(os.environ.get("HF_HUB_DISABLE_SHARED_BLOBS"))
 
 # Bucket hosting the static sandbox server binary (see huggingface_hub.Sandbox)
